告警修改

+++ services/implementation/src/authentication_v2/auth_stages/auth_credential.cpp
@@ -157,7 +157,7 @@ int32_t AuthSrcCredentialAuthDoneState::Action(std::shared_ptr<DmAuthContext> co
 
     // first time joinLnn, auth lnnCredential
     if (context->accesser.isGenerateLnnCredential == true && context->isAppCredentialVerified == false &&
-        context->accesser.bindLevel != USER) {
+        context->accesser.bindLevel != static_cast<int32_t>(USER)) {
         context->isAppCredentialVerified = true;
         DerivativeSessionKey(context);
         msgType = MSG_TYPE_REQ_CREDENTIAL_AUTH_START;
@@ -174,7 +174,8 @@ int32_t AuthSrcCredentialAuthDoneState::Action(std::shared_ptr<DmAuthContext> co
             return ERR_DM_FAILED;
         }
         // First-time authentication and Lnn credential process
-    } else if (context->accesser.isGenerateLnnCredential == true && context->accesser.bindLevel != USER) {
+    } else if (context->accesser.isGenerateLnnCredential == true &&
+        context->accesser.bindLevel != static_cast<int32_t>(USER)) {
         int32_t skId = 0;
         int32_t ret = context->authMessageProcessor->SaveSessionKeyToDP(context->accesser.userId, skId);
         if (ret != DM_OK) {
@@ -315,7 +316,8 @@ int32_t AuthSinkCredentialAuthNegotiateState::Action(std::shared_ptr<DmAuthConte
     int32_t skId;
 
     // First lnn cred auth, second time receiving 161 message
-    if (context->accessee.isGenerateLnnCredential == true && context->accessee.bindLevel != USER &&
+    if (context->accessee.isGenerateLnnCredential == true &&
+        context->accessee.bindLevel != static_cast<int32_t>(USER) &&
         context->isAppCredentialVerified == true) {
         ret = context->authMessageProcessor->SaveSessionKeyToDP(context->accessee.userId, skId);
         if (ret != DM_OK) {
@@ -546,7 +548,7 @@ int32_t AuthSrcCredentialExchangeState::Action(std::shared_ptr<DmAuthContext> co
         return DM_OK;
     }
     // First authentication, generate LNN credentials and public key
-    if (context->accesser.isGenerateLnnCredential && context->accesser.bindLevel != USER) {
+    if (context->accesser.isGenerateLnnCredential && context->accesser.bindLevel != static_cast<int32_t>(USER)) {
         ret = GenerateCredIdAndPublicKey(DM_AUTH_SCOPE_LNN, context);
         if (ret != DM_OK) {
             LOGE("AuthSrcCredentialExchangeState::Action() error, generate user credId and publicKey failed.");
@@ -555,9 +557,10 @@ int32_t AuthSrcCredentialExchangeState::Action(std::shared_ptr<DmAuthContext> co
     }
 
     DmAuthScope authorizedScope = DM_AUTH_SCOPE_INVALID;
-    if (context->accesser.bindLevel == APP || context->accesser.bindLevel == SERVICE) {
+    if (context->accesser.bindLevel == static_cast<int32_t>(APP) ||
+        context->accesser.bindLevel == static_cast<int32_t>(SERVICE)) {
         authorizedScope = DM_AUTH_SCOPE_APP;
-    } else if (context->accesser.bindLevel == USER) {
+    } else if (context->accesser.bindLevel == static_cast<int32_t>(USER)) {
         authorizedScope = DM_AUTH_SCOPE_USER;
     }
 
@@ -591,7 +594,7 @@ int32_t AuthSinkCredentialExchangeState::Action(std::shared_ptr<DmAuthContext> c
     }
 
     // First authentication lnn cred
-    if (context->accessee.isGenerateLnnCredential && context->accessee.bindLevel != USER) {
+    if (context->accessee.isGenerateLnnCredential && context->accessee.bindLevel != static_cast<int32_t>(USER)) {
         // Generate credentials and public key
         ret = GenerateCredIdAndPublicKey(DM_AUTH_SCOPE_LNN, context);
         if (ret != DM_OK) {
@@ -614,9 +617,10 @@ int32_t AuthSinkCredentialExchangeState::Action(std::shared_ptr<DmAuthContext> c
     }
 
     DmAuthScope authorizedScope = DM_AUTH_SCOPE_INVALID;
-    if (context->accessee.bindLevel == APP || context->accessee.bindLevel == SERVICE) {
+    if (context->accessee.bindLevel == static_cast<int32_t>(APP) ||
+        context->accessee.bindLevel == static_cast<int32_t>(SERVICE)) {
         authorizedScope = DM_AUTH_SCOPE_APP;
-    } else if (context->accessee.bindLevel == USER) {
+    } else if (context->accessee.bindLevel == static_cast<int32_t>(USER)) {
         authorizedScope = DM_AUTH_SCOPE_USER;
     }
     // Generate transport credentials and public key
@@ -663,7 +667,7 @@ int32_t AuthSrcCredentialAuthStartState::Action(std::shared_ptr<DmAuthContext> c
 
     if (IsNeedAgreeCredential(context)) {
         // First authentication
-        if (context->accesser.isGenerateLnnCredential && context->accesser.bindLevel != USER) {
+        if (context->accesser.isGenerateLnnCredential && context->accesser.bindLevel != static_cast<int32_t>(USER)) {
             // Agree lnn credentials and public key
             tmpCredId = context->accesser.lnnCredentialId;
             ret = AgreeCredential(DM_AUTH_SCOPE_LNN, context);
@@ -678,9 +682,10 @@ int32_t AuthSrcCredentialAuthStartState::Action(std::shared_ptr<DmAuthContext> c
         }
 
         DmAuthScope authorizedScope = DM_AUTH_SCOPE_INVALID;
-        if (context->accesser.bindLevel == APP || context->accesser.bindLevel == SERVICE) {
+        if (context->accesser.bindLevel == static_cast<int32_t>(APP) ||
+            context->accesser.bindLevel == static_cast<int32_t>(SERVICE)) {
             authorizedScope = DM_AUTH_SCOPE_APP;
-        } else if (context->accesser.bindLevel == USER) {
+        } else if (context->accesser.bindLevel == static_cast<int32_t>(USER)) {
             authorizedScope = DM_AUTH_SCOPE_USER;
         }
